fix #2675
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@13355 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/app/qgsmaptoolnodetool.cpp
@@ -864,7 +864,7 @@ void QgsMapToolNodeTool::keyReleaseEvent( QKeyEvent* e )
     mCtrl = false;
     return;
   }
-  if ( e->key() == Qt::Key_Delete )
+  if ( mSelectionFeature && e->key() == Qt::Key_Delete )
   {
     mChangingGeometry = true;
     mSelectionFeature->deleteSelectedVertexes();
